+++ Code/core/Utility.h
@@ -22,7 +22,7 @@ public:
 	static double Distance(CvPoint2D32f p1, CvPoint2D32f p2) ;
 
 	static wxString toMillisString(wxDateTime ts);
-	static float toMillis(wxDateTime ts);
+	static double toMillis(wxDateTime ts);
 
 	//! Creates a rectangle from center, width and height.
 	static CvRect RectByCenter(int center_x, int center_xy, int width, int height);
